Stop treating linked libraries as in-project files
Currently files from linked libraries are synced just like normal
project files. This causes the same library file to exist multiple times
on the local disk (in every project it is linked to). If one of the
local versions is updated, this update is currently not propagated to
the other local versions of the file.

To fix this the CLI will stop treating files from linked libraries as
in-project files. This means that only the original versions (the
version of the file in its library project in the 'Library' directory)
will be synced. By working this way, every library file only exists once
on the local disk (in its project in the 'Library' directory).

If a user tries to push, pull or delete a library file in a non-library
project a warning will be shown letting him know of the above. Users
can safely remove local versions of library files in non-library
projects.

Because this change makes it impossible to know which project has access
to which files in a QuantConnect CLI project (there is no way to see
which libraries are linked to a project anymore) a projects:info command
will be added soon which will display this information.

+++ src/commands/files/watch.ts
@@ -68,6 +68,15 @@ export default class WatchCommand extends BaseCommand {
       return;
     }
 
+    const isLibrary = await this.isLibraryFile(projectId, relativePath);
+    if (isLibrary) {
+      logger.warn('QuantConnect CLI no longer treats files from linked libraries as in-project files');
+      logger.warn(`For that reason '${file}' was not pushed`);
+      logger.warn('You can safely remove this file from your local disk');
+      logger.warn("In the future please edit this file in its project in the 'Library' directory");
+      return;
+    }
+
     try {
       await this.api.files.update(projectId, relativePath, fileContent);
       logger.info(`Successfully pushed update to '${file}'`);
@@ -91,6 +100,16 @@ export default class WatchCommand extends BaseCommand {
     const absolutePath = path.resolve(process.cwd(), file);
     const relativePath = getFilePathRelativeToProject(projectName, absolutePath);
 
+    const isLibrary = await this.isLibraryFile(projectId, relativePath);
+    if (isLibrary) {
+      logger.warn('QuantConnect CLI no longer treats files from linked libraries as in-project files');
+      logger.warn(`For that reason the deletion of '${file}' was not pushed`);
+      logger.warn(
+        "Please remove this file in its project in the 'Library' directory if you want it deleted from the library",
+      );
+      return;
+    }
+
     try {
       await this.api.files.delete(projectId, relativePath);
       logger.info(`Successfully pushed deletion of '${file}'`);
@@ -105,5 +124,10 @@ export default class WatchCommand extends BaseCommand {
     }
 
     return fs.existsSync(file) ? fs.statSync(file).isDirectory() : path.extname(file) === '';
+  }
+
+  private async isLibraryFile(projectId: number, relativePath: string): Promise<boolean> {
+    const remoteFiles = await this.api.files.getAll(projectId);
+    return remoteFiles.some(file => file.isLibrary && file.name === relativePath);
   }
 }
